# Improve the performance of method CandidatesToVictimsMap